requirements/requirements.txt
absl-py>=0.9.0
chex>=0.0.8
distrax>=0.0.2
dm_env
jax>=0.1.55
jaxlib>=0.1.37
numpy>=1.18.0
